requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.2.5
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djongo==1.3.6
idna==3.2
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
requests==2.26.0
simplejson==3.17.3
sqlparse==0.2.4
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
